add debugging settings for exception on loadingDialog close

+++ Twitch_Viewer/Types/StreamItem.cs
@@ -84,7 +84,10 @@ namespace Twitch_Viewer.Types
 
             loadingDialog.Closing += (dSender, dArgs) =>
             {
-                if (p?.HasExited != true)
+                if (p == null && !MainWindow.settings.DisableNullCheck)
+                    throw new Exception("process object is null. Error might be here.");
+
+                if (p?.HasExited != true && !MainWindow.settings.DisableProcessKill)
                     p?.Kill();
             };
